requirements.txt
codegen==1.0
Django==1.10.5
djangorestframework==3.5.4
MySQL-python==1.2.5
patterns==0.3
Pillow==5.0.0
PyMySQL==0.8.0
uWSGI==2.0.17
